airlock::config

Struct ADPConfig

pub struct ADPConfig {
    pub image: String,
    pub entrypoint: Vec<String>,
    pub command: Vec<String>,
    pub config_path: PathBuf,
    pub additional_env_args: Vec<String>,
}

Agent Data Plane configuration.

Fields
image: String

Container image to use.

This must be a valid image reference โ agent-data-plane:x.y.z, registry.ddbuild.io/saluki/agent-data-plane:x.y.z, etc โ to an image containing the agent-data-plane binary.

The agent-data-plane binary must exist at /usr/bin/agent-data-plane in the image. Otherwise, the binary path can be overridden with the binary-path argument.

entrypoint: Vec<String>

Entrypoint to execute.

command: Vec<String>

Command to run Agent Data Plane with.

config_path: PathBuf

Path to the Agent Data Plane configuration file to use.

This file is mapped into the ADP container and so it must exist on the system where this command is run from.

additional_env_args: Vec<String>

Additional environment variables to be passed into the Agent Data Plane container.

These should be in the form of KEY=VALUE.